refactored Primitive type names
