+++ Cargo.toml
@@ -10,7 +10,7 @@ bytemuck = { version = "1.0", features = [
 	"zeroable_atomics",
 ] }
 iced = { version = "0.12.0", features = ["debug", "advanced", "svg"] }
-iced_aw = { version = "0.8.0", default-features = false, features = [
+iced_aw = { version = "0.9.2", default-features = false, features = [
 	"color_picker",
 	"card",
 	"modal",
