+++ src/render/pipeline/buffer/types.rs
@@ -1,6 +1,15 @@
 use iced::Color;
 use ultraviolet::{Mat4, Vec4};
 
+#[derive(Debug, Copy, Clone, bytemuck::Pod, bytemuck::Zeroable)]
+#[repr(C)]
+pub struct Vertex {
+    pub normal: Vec4,
+    pub barycentric: Vec4,
+    pub sides: Vec4,
+    pub color: Vec4,
+}
+
 pub struct AllUniforms {
     pub model: ModelUniforms,
     pub frag: FragUniforms,
